pyproject.toml
[build-system]
requires = ["setuptools", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "npr-cli"
version = "0.0.2"
authors = [
    { name="evanmags", email="magnussen.evan@gmail.com" }
]
description = "A simple npr streaming cli"
readme = "README.md"
dependencies = [
    "click",
    "inquirerpy",
    "python-vlc",
    "requests",
]
requires-python = ">=3.10"
classifiers = [
    "Environment :: Console",
    "License :: OSI Approved :: MIT License",
    "Operating System :: MacOS",
    "Programming Language :: Python :: 3.10",
    "Topic :: Multimedia :: Sound/Audio :: Players",
]

[project.urls]

[project.scripts]
npr = "npr.cli:npr"

[tool.setuptools.packages.find]
where = ["src"]